refactor: API embedding by modules
